restructured schema generation
